requirements.txt
pytest>=7.0
requests>=2.28
python-dotenv>=1.0
pytest-check>=0.4.0